Part 2: Test Driven Design (TDD) for trades from Restful API
RestfulTradeDataProviderTests was passed and the program worked properly.

Request: You decide to provide the option to read trades from a Restful web API

Request 409 - "As a trader, I want to be able to able to read trades from the companies new restful web api so I can enter trades from different apps."
TDD: Get the RestfulTradeDataProviderTests to pass
Details: You must modify the RestfulTradeDataProvider class and make it work, reading a file from the url passed in as a parameter to the constructor.
Try using AI to help with to modify the GetTradeData() method that supplies the data from the service. You should to :
Use HttpClient and GetAsync to connect to the server
Use ReadAsStringAsync() to read the data from
Use JsonSerializer to convert the Json data into a trade string.  The Json file looks like this: https://unit9trader.azurewebsites.net/api/TradeData
This will require your code run in a sperate, asynchronous task
Server: The Restful API server is already running in the cloud. Here is the code for the Rest API provider.