Part 3: Unit test of trade validation
All SimpleTradeValidatorTests were passed

Request: The validation rules of the input data change.

Request 403 - "As a Traders, I want to prevent dangerous trades so that one typo doesn't ruin the company."
Details: Trade amounts must be between 1,000 and 1,000,000 units. If the trade is outside these bounds it should be ignored.
Solution: Edit the ValidateTradeData method.
Tests: All the test for ValidateTradeData should pass.